+++ packages/firebase_auth/android/src/main/java/io/flutter/plugins/firebaseauth/FirebaseAuthPlugin.java
@@ -16,6 +16,7 @@ import com.google.firebase.FirebaseException;
 import com.google.firebase.FirebaseNetworkException;
 import com.google.firebase.FirebaseTooManyRequestsException;
 import com.google.firebase.auth.ActionCodeSettings;
+import com.google.firebase.auth.AdditionalUserInfo;
 import com.google.firebase.auth.AuthCredential;
 import com.google.firebase.auth.AuthResult;
 import com.google.firebase.auth.EmailAuthProvider;
@@ -36,6 +37,7 @@ import com.google.firebase.auth.SignInMethodQueryResult;
 import com.google.firebase.auth.TwitterAuthProvider;
 import com.google.firebase.auth.UserInfo;
 import com.google.firebase.auth.UserProfileChangeRequest;
+import com.google.gson.Gson;
 import io.flutter.plugin.common.MethodCall;
 import io.flutter.plugin.common.MethodChannel;
 import io.flutter.plugin.common.MethodChannel.MethodCallHandler;
@@ -138,6 +140,9 @@ public class FirebaseAuthPlugin implements MethodCallHandler {
       case "updateEmail":
         handleUpdateEmail(call, result, getAuth(call));
         break;
+      case "updatePhoneNumberCredential":
+        handleUpdatePhoneNumber(call, result, getAuth(call));
+        break;
       case "updatePassword":
         handleUpdatePassword(call, result, getAuth(call));
         break;
@@ -190,37 +195,11 @@ public class FirebaseAuthPlugin implements MethodCallHandler {
         new PhoneAuthProvider.OnVerificationStateChangedCallbacks() {
           @Override
           public void onVerificationCompleted(PhoneAuthCredential phoneAuthCredential) {
-            if (link) {
-              firebaseAuth
-                  .getCurrentUser()
-                  .linkWithCredential(phoneAuthCredential)
-                  .addOnCompleteListener(
-                      new OnCompleteListener<AuthResult>() {
-
-                        @Override
-                        public void onComplete(@NonNull Task<AuthResult> task) {
-                          if (task.isSuccessful()) {
-                            Map<String, Object> arguments = new HashMap<>();
-                            arguments.put("handle", handle);
-                            channel.invokeMethod("phoneVerificationCompleted", arguments);
-                          }
-                        }
-                      });
-            } else {
-              firebaseAuth
-                  .signInWithCredential(phoneAuthCredential)
-                  .addOnCompleteListener(
-                      new OnCompleteListener<AuthResult>() {
-                        @Override
-                        public void onComplete(@NonNull Task<AuthResult> task) {
-                          if (task.isSuccessful()) {
-                            Map<String, Object> arguments = new HashMap<>();
-                            arguments.put("handle", handle);
-                            channel.invokeMethod("phoneVerificationCompleted", arguments);
-                          }
-                        }
-                      });
-            }
+            Map<String, Object> arguments = new HashMap<>();
+            arguments.put("handle", handle);
+            String parsedJson = new Gson().toJson(phoneAuthCredential);
+            arguments.put("phoneAuthCredential", parsedJson);
+            channel.invokeMethod("phoneVerificationCompleted", arguments);
           }
 
           @Override
@@ -473,9 +452,13 @@ public class FirebaseAuthPlugin implements MethodCallHandler {
         }
       case PhoneAuthProvider.PROVIDER_ID:
         {
-          String verificationId = data.get("verificationId");
-          String smsCode = data.get("smsCode");
-          credential = PhoneAuthProvider.getCredential(verificationId, smsCode);
+          if (data.containsKey("verificationId")) {
+            String accessToken = data.get("verificationId");
+            String smsCode = data.get("smsCode");
+            credential = PhoneAuthProvider.getCredential(accessToken, smsCode);
+          } else {
+            credential = new Gson().fromJson(data.get("jsonObject"), PhoneAuthCredential.class);
+          }
           break;
         }
       default:
@@ -551,7 +534,7 @@ public class FirebaseAuthPlugin implements MethodCallHandler {
     }
 
     Map<String, Boolean> arguments = call.arguments();
-    boolean refresh = arguments.get("refresh");
+    final boolean refresh = arguments.get("refresh");
 
     currentUser
         .getIdToken(refresh)
@@ -559,8 +542,18 @@ public class FirebaseAuthPlugin implements MethodCallHandler {
             new OnCompleteListener<GetTokenResult>() {
               public void onComplete(@NonNull Task<GetTokenResult> task) {
                 if (task.isSuccessful() && task.getResult() != null) {
-                  String idToken = task.getResult().getToken();
-                  result.success(idToken);
+                  final Map<String, Object> map = new HashMap<>();
+                  map.put("token", task.getResult().getToken());
+                  map.put("expirationTimestamp", task.getResult().getExpirationTimestamp());
+                  map.put("authTimestamp", task.getResult().getAuthTimestamp());
+                  map.put("issuedAtTimestamp", task.getResult().getIssuedAtTimestamp());
+                  map.put("claims", task.getResult().getClaims());
+
+                  if (task.getResult().getSignInProvider() != null) {
+                    map.put("signInProvider", task.getResult().getSignInProvider());
+                  }
+
+                  result.success(Collections.unmodifiableMap(map));
                 } else {
                   reportException(result, task.getException());
                 }
@@ -579,6 +572,16 @@ public class FirebaseAuthPlugin implements MethodCallHandler {
     final String email = arguments.get("email");
 
     currentUser.updateEmail(email).addOnCompleteListener(new TaskVoidCompleteListener(result));
+  }
+
+  private void handleUpdatePhoneNumber(MethodCall call, Result result, FirebaseAuth firebaseAuth) {
+    @SuppressWarnings("unchecked")
+    AuthCredential credential = getCredential((Map<String, Object>) call.arguments);
+
+    firebaseAuth
+        .getCurrentUser()
+        .updatePhoneNumber((PhoneAuthCredential) credential)
+        .addOnCompleteListener(new TaskVoidCompleteListener(result));
   }
 
   private void handleUpdatePassword(MethodCall call, Result result, FirebaseAuth firebaseAuth) {
@@ -679,9 +682,15 @@ public class FirebaseAuthPlugin implements MethodCallHandler {
       if (!task.isSuccessful() || task.getResult() == null) {
         reportException(result, task.getException());
       } else {
-        FirebaseUser user = task.getResult().getUser();
-        Map<String, Object> userMap = Collections.unmodifiableMap(mapFromUser(user));
-        result.success(userMap);
+        AuthResult authResult = task.getResult();
+        FirebaseUser user = authResult.getUser();
+        AdditionalUserInfo additionalUserInfo = authResult.getAdditionalUserInfo();
+        Map<String, Object> userMap = (mapFromUser(user));
+        Map<String, Object> additionalUserInfoMap = mapFromAdditionalUserInfo(additionalUserInfo);
+        Map<String, Object> map = new HashMap<>();
+        map.put("user", userMap);
+        map.put("additionalUserInfo", additionalUserInfoMap);
+        result.success(Collections.unmodifiableMap(map));
       }
     }
   }
@@ -745,11 +754,6 @@ public class FirebaseAuthPlugin implements MethodCallHandler {
     if (user != null) {
       List<Map<String, Object>> providerData = new ArrayList<>();
       for (UserInfo userInfo : user.getProviderData()) {
-        // Ignore phone provider since firebase provider is a super set of the phone
-        // provider.
-        if (userInfo.getProviderId().equals("phone")) {
-          continue;
-        }
         providerData.add(Collections.unmodifiableMap(userInfoToMap(userInfo)));
       }
       Map<String, Object> userMap = userInfoToMap(user);
@@ -762,6 +766,19 @@ public class FirebaseAuthPlugin implements MethodCallHandler {
       userMap.put("isEmailVerified", user.isEmailVerified());
       userMap.put("providerData", Collections.unmodifiableList(providerData));
       return Collections.unmodifiableMap(userMap);
+    } else {
+      return null;
+    }
+  }
+
+  private Map<String, Object> mapFromAdditionalUserInfo(AdditionalUserInfo info) {
+    if (info != null) {
+      Map<String, Object> additionalUserInfoMap = new HashMap<>();
+      additionalUserInfoMap.put("profile", info.getProfile());
+      additionalUserInfoMap.put("providerId", info.getProviderId());
+      additionalUserInfoMap.put("username", info.getUsername());
+      additionalUserInfoMap.put("isNewUser", info.isNewUser());
+      return Collections.unmodifiableMap(additionalUserInfoMap);
     } else {
       return null;
     }