+++ packages/firebase_auth/example/lib/signin_page.dart
@@ -1,3 +1,7 @@
+// Copyright 2019 The Chromium Authors. All rights reserved.
+// Use of this source code is governed by a BSD-style license that can be
+// found in the LICENSE file.
+
 import 'dart:io';
 import 'package:flutter/material.dart';
 import 'package:firebase_auth/firebase_auth.dart';
@@ -94,6 +98,7 @@ class _EmailPasswordFormState extends State<_EmailPasswordForm> {
               if (value.isEmpty) {
                 return 'Please enter some text';
               }
+              return null;
             },
           ),
           TextFormField(
@@ -103,6 +108,7 @@ class _EmailPasswordFormState extends State<_EmailPasswordForm> {
               if (value.isEmpty) {
                 return 'Please enter some text';
               }
+              return null;
             },
           ),
           Container(
@@ -143,10 +149,11 @@ class _EmailPasswordFormState extends State<_EmailPasswordForm> {
 
   // Example code of how to sign in with email and password.
   void _signInWithEmailAndPassword() async {
-    final FirebaseUser user = await _auth.signInWithEmailAndPassword(
+    final FirebaseUser user = (await _auth.signInWithEmailAndPassword(
       email: _emailController.text,
       password: _passwordController.text,
-    );
+    ))
+        .user;
     if (user != null) {
       setState(() {
         _success = true;
@@ -191,10 +198,11 @@ class _EmailLinkSignInSectionState extends State<_EmailLinkSignInSection>
       final Uri link = await _retrieveDynamicLink();
 
       if (link != null) {
-        final FirebaseUser user = await _auth.signInWithEmailAndLink(
+        final FirebaseUser user = (await _auth.signInWithEmailAndLink(
           email: _userEmail,
           link: link.toString(),
-        );
+        ))
+            .user;
 
         if (user != null) {
           _userID = user.uid;
@@ -235,6 +243,7 @@ class _EmailLinkSignInSectionState extends State<_EmailLinkSignInSection>
               if (value.isEmpty) {
                 return 'Please enter your email.';
               }
+              return null;
             },
           ),
           Container(
@@ -327,7 +336,7 @@ class _AnonymouslySignInSectionState extends State<_AnonymouslySignInSection> {
 
   // Example code of how to sign in anonymously.
   void _signInAnonymously() async {
-    final FirebaseUser user = await _auth.signInAnonymously();
+    final FirebaseUser user = (await _auth.signInAnonymously()).user;
     assert(user != null);
     assert(user.isAnonymous);
     assert(!user.isEmailVerified);
@@ -410,7 +419,8 @@ class _GoogleSignInSectionState extends State<_GoogleSignInSection> {
       accessToken: googleAuth.accessToken,
       idToken: googleAuth.idToken,
     );
-    final FirebaseUser user = await _auth.signInWithCredential(credential);
+    final FirebaseUser user =
+        (await _auth.signInWithCredential(credential)).user;
     assert(user.email != null);
     assert(user.displayName != null);
     assert(!user.isAnonymous);
@@ -462,6 +472,7 @@ class _PhoneSignInSectionState extends State<_PhoneSignInSection> {
             if (value.isEmpty) {
               return 'Phone number (+x xxx-xxx-xxxx)';
             }
+            return null;
           },
         ),
         Container(
@@ -500,15 +511,16 @@ class _PhoneSignInSectionState extends State<_PhoneSignInSection> {
     );
   }
 
-  // Exmaple code of how to veify phone number
+  // Example code of how to verify phone number
   void _verifyPhoneNumber() async {
     setState(() {
       _message = '';
     });
     final PhoneVerificationCompleted verificationCompleted =
-        (FirebaseUser user) {
+        (AuthCredential phoneAuthCredential) {
+      _auth.signInWithCredential(phoneAuthCredential);
       setState(() {
-        _message = 'signInWithPhoneNumber auto succeeded: $user';
+        _message = 'Received phone auth credential: $phoneAuthCredential';
       });
     };
 
@@ -549,7 +561,8 @@ class _PhoneSignInSectionState extends State<_PhoneSignInSection> {
       verificationId: _verificationId,
       smsCode: _smsController.text,
     );
-    final FirebaseUser user = await _auth.signInWithCredential(credential);
+    final FirebaseUser user =
+        (await _auth.signInWithCredential(credential)).user;
     final FirebaseUser currentUser = await _auth.currentUser();
     assert(user.uid == currentUser.uid);
     setState(() {
@@ -693,7 +706,8 @@ class _OtherProvidersSignInSectionState
     final AuthCredential credential = GithubAuthProvider.getCredential(
       token: _tokenController.text,
     );
-    final FirebaseUser user = await _auth.signInWithCredential(credential);
+    final FirebaseUser user =
+        (await _auth.signInWithCredential(credential)).user;
     assert(user.email != null);
     assert(user.displayName != null);
     assert(!user.isAnonymous);
@@ -715,7 +729,8 @@ class _OtherProvidersSignInSectionState
     final AuthCredential credential = FacebookAuthProvider.getCredential(
       accessToken: _tokenController.text,
     );
-    final FirebaseUser user = await _auth.signInWithCredential(credential);
+    final FirebaseUser user =
+        (await _auth.signInWithCredential(credential)).user;
     assert(user.email != null);
     assert(user.displayName != null);
     assert(!user.isAnonymous);
@@ -737,7 +752,8 @@ class _OtherProvidersSignInSectionState
     final AuthCredential credential = TwitterAuthProvider.getCredential(
         authToken: _tokenController.text,
         authTokenSecret: _tokenSecretController.text);
-    final FirebaseUser user = await _auth.signInWithCredential(credential);
+    final FirebaseUser user =
+        (await _auth.signInWithCredential(credential)).user;
     assert(user.email != null);
     assert(user.displayName != null);
     assert(!user.isAnonymous);