+++ packages/firebase_remote_config/android/src/main/java/io/flutter/plugins/firebase/firebaseremoteconfig/FirebaseRemoteConfigPlugin.java
@@ -1,3 +1,7 @@
+// Copyright 2019 The Chromium Authors. All rights reserved.
+// Use of this source code is governed by a BSD-style license that can be
+// found in the LICENSE file.
+
 package io.flutter.plugins.firebase.firebaseremoteconfig;
 
 import android.content.Context;
@@ -29,17 +33,12 @@ public class FirebaseRemoteConfigPlugin implements MethodCallHandler {
   public static final String DEFAULT_PREF_KEY = "default_keys";
 
   private static SharedPreferences sharedPreferences;
-  private final MethodChannel channel;
 
   public static void registerWith(Registrar registrar) {
     final MethodChannel channel =
         new MethodChannel(registrar.messenger(), "plugins.flutter.io/firebase_remote_config");
-    channel.setMethodCallHandler(new FirebaseRemoteConfigPlugin(channel));
+    channel.setMethodCallHandler(new FirebaseRemoteConfigPlugin());
     sharedPreferences = registrar.context().getSharedPreferences(PREFS_NAME, Context.MODE_PRIVATE);
-  }
-
-  private FirebaseRemoteConfigPlugin(MethodChannel channel) {
-    this.channel = channel;
   }
 
   @Override